+++ android/src/main/kotlin/id/passage/passage_flutter/PassageFlutter.kt
@@ -6,12 +6,14 @@ import com.google.gson.Gson
 import id.passage.android.Passage
 import id.passage.android.PassageSocialConnection
 import id.passage.android.PassageToken
+import id.passage.android.PasskeyCreationOptions
 import id.passage.android.exceptions.AppInfoException
 import id.passage.android.exceptions.LoginWithPasskeyCancellationException
 import id.passage.android.exceptions.OneTimePasscodeActivateExceededAttemptsException
 import id.passage.android.exceptions.PassageUserException
 import id.passage.android.exceptions.PassageUserUnauthorizedException
 import id.passage.android.exceptions.RegisterWithPasskeyCancellationException
+import id.passage.android.model.AuthenticatorAttachment
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import kotlinx.coroutines.CoroutineScope
@@ -34,12 +36,20 @@ internal class PassageFlutter(private val activity: Activity, appId: String? = n
 
     // region PASSKEY METHODS
 
-    internal fun register(call: MethodCall, result: MethodChannel.Result) {
+    internal fun registerWithPasskey(call: MethodCall, result: MethodChannel.Result) {
         val identifier = call.argument<String>("identifier")
             ?: return invalidArgumentError(result)
         CoroutineScope(Dispatchers.IO).launch {
             try {
-                val authResult = passage.registerWithPasskey(identifier)
+                var options: PasskeyCreationOptions? = null
+                call.argument<Map<String, String>>("options")?.let { map ->
+                    map["authenticatorAttachment"]?.let { authAttachmentString ->
+                        AuthenticatorAttachment.decode(authAttachmentString)?.let { authenticatorAttachment ->
+                            options = PasskeyCreationOptions(authenticatorAttachment)
+                        }
+                    }
+                }
+                val authResult = passage.registerWithPasskey(identifier, options)
                 val jsonString = Gson().toJson(authResult)
                 result.success(jsonString)
             } catch (e: Exception) {
@@ -54,10 +64,11 @@ internal class PassageFlutter(private val activity: Activity, appId: String? = n
         }
     }
 
-    internal fun login(result: MethodChannel.Result) {
+    internal fun loginWithPasskey(call: MethodCall, result: MethodChannel.Result) {
+        val identifier = call.argument<String>("identifier")
         CoroutineScope(Dispatchers.IO).launch {
             try {
-                val authResult = passage.loginWithPasskey("")
+                val authResult = passage.loginWithPasskey(identifier)
                 val jsonString = Gson().toJson(authResult)
                 result.success(jsonString)
             } catch (e: Exception) {
@@ -299,11 +310,20 @@ internal class PassageFlutter(private val activity: Activity, appId: String? = n
         }
     }
 
-    fun addPasskey(result: MethodChannel.Result) {
+    fun addPasskey(call: MethodCall, result: MethodChannel.Result) {
         CoroutineScope(Dispatchers.IO).launch {
             try {
-                val user = passage.getCurrentUser() ?: throw PassageUserUnauthorizedException("User is not authorized.")
-                val credential = user.addDevicePasskey(activity)
+                val user = passage.getCurrentUser()
+                    ?: throw PassageUserUnauthorizedException("User is not authorized.")
+                var options: PasskeyCreationOptions? = null
+                call.argument<Map<String, String>>("options")?.let { map ->
+                    map["authenticatorAttachment"]?.let { authAttachmentString ->
+                        AuthenticatorAttachment.decode(authAttachmentString)?.let { authenticatorAttachment ->
+                            options = PasskeyCreationOptions(authenticatorAttachment)
+                        }
+                    }
+                }
+                val credential = user.addDevicePasskey(activity, options)
                 val jsonString = Gson().toJson(credential)
                 result.success(jsonString)
             } catch (e: Exception) {